Corrected H264/Nalu length to unsigned for from-fourcc funcs

+++ src/SoyAvf.h
@@ -7,7 +7,7 @@
 namespace Avf
 {
 	class TAsset;
-	SoyMediaFormat::Type			SoyMediaFormat_FromFourcc(uint32 Fourcc,int H264LengthSize);
+	SoyMediaFormat::Type			SoyMediaFormat_FromFourcc(uint32 Fourcc,size_t H264LengthSize);
 	void							GetFileExtensions(ArrayBridge<std::string>&& Extensions);
 
 #if defined(__OBJC__)
